+++ bluemira/geometry/coordinates.py
@@ -35,7 +35,7 @@ from scipy.spatial.distance import cdist
 
 from bluemira.base.constants import EPS
 from bluemira.base.look_and_feel import bluemira_warn
-from bluemira.geometry.constants import CROSS_P_TOL
+from bluemira.geometry.constants import CROSS_P_TOL, DOT_P_TOL
 from bluemira.geometry.error import CoordinatesError
 from bluemira.utilities.tools import json_writer
 
@@ -624,6 +624,241 @@ def principal_components(xyz_array):
     eigenvalues = eigenvalues[sort]
     eigenvectors = eigenvectors[:, sort]
     return eigenvalues, eigenvectors
+
+
+# =============================================================================
+# Intersection tools
+# =============================================================================
+
+
+def vector_intersect_3d(p_1, p_2, p_3, p_4):
+    """
+    Get the intersection point between two 3-D vectors.
+
+    Parameters
+    ----------
+    p1: np.ndarray(3)
+        The first point on the first vector
+    p2: np.ndarray(3)
+        The second point on the first vector
+    p3: np.ndarray(3)
+        The first point on the second vector
+    p4: np.ndarray(3)
+        The second point on the second vector
+
+    Returns
+    -------
+    p_inter: np.ndarray(3)
+        The point of the intersection between the two vectors
+
+    Raises
+    ------
+    CoordinatesError
+        If there is no intersection between the points
+
+    Notes
+    -----
+    Credit: Paul Bourke at
+    http://paulbourke.net/geometry/pointlineplane/#:~:text=The%20shortest%20line%20between%20two%20lines%20in%203D
+    """
+    p_13 = p_1 - p_3
+    p_43 = p_4 - p_3
+
+    if np.linalg.norm(p_13) < EPS:
+        raise CoordinatesError("No intersection between 3-D lines.")
+    p_21 = p_2 - p_1
+    if np.linalg.norm(p_21) < EPS:
+        raise CoordinatesError("No intersection between 3-D lines.")
+
+    d1343 = np.dot(p_13, p_43)
+    d4321 = np.dot(p_43, p_21)
+    d1321 = np.dot(p_13, p_21)
+    d4343 = np.dot(p_43, p_43)
+    d2121 = np.dot(p_21, p_21)
+
+    denom = d2121 * d4343 - d4321 * d4321
+
+    if np.abs(denom) < EPS:
+        raise CoordinatesError("No intersection between 3-D lines.")
+
+    numer = d1343 * d4321 - d1321 * d4343
+
+    mua = numer / denom
+    intersection = p_1 + mua * p_21
+    return intersection
+
+
+def coords_plane_intersect(coords, plane):
+    """
+    Calculate the intersection of Coordinates with a plane.
+
+    Parameters
+    ----------
+    coords: Coordinates
+        The coordinates to calculate the intersection on
+    plane: Plane
+        The plane to calculate the intersection with
+
+    Returns
+    -------
+    inter: np.array(3, n_intersections) or None
+        The xyz coordinates of the intersections with the Coordinates. Returns None if
+        there are no intersections detected
+    """
+    out = _coords_plane_intersect(coords.xyz.T[:-1], plane.base, plane.axis)
+    if not out:
+        return None
+    else:
+        return np.unique(out, axis=0)  # Drop occasional duplicates
+
+
+@nb.jit(cache=True, nopython=True)
+def _coords_plane_intersect(array, p1, vec2):
+    # JIT compiled utility of the above
+    out = []
+    for i in range(len(array)):
+        vec1 = array[i + 1] - array[i]
+        dot = np.dot(vec1, vec2)
+        if abs(dot) > DOT_P_TOL:
+            w = array[i] - p1
+            fac = -(np.dot(vec2, w)) / dot
+            if (fac >= 0) and (fac <= 1):
+                out.append(array[i] + fac * vec1)
+    return out
+
+
+def get_intersect(xy1, xy2):
+    """
+    Calculates the intersection points between two sets of 2-D coordinates. Will return
+    a unique list of x, z intersections (no duplicates in x-z space).
+
+    Parameters
+    ----------
+    xy1: np.ndarray
+        The 2-D coordinates between which intersection points should be calculated
+    xy2: np.ndarray
+        The 2-D coordinates between which intersection points should be calculated
+
+    Returns
+    -------
+    xi: np.array(N_itersection)
+        The x coordinates of the intersection points
+    zi: np.array(N_itersection)
+        The z coordinates of the intersection points#
+
+    Note
+    ----
+    D. Schwarz, <https://uk.mathworks.com/matlabcentral/fileexchange/11837-fast-and-robust-curve-intersections>
+    """  # noqa :W505
+    x1, y1 = xy1
+    x2, y2 = xy2
+
+    def inner_inter(x_1, x_2):
+        n1, n2 = x_1.shape[0] - 1, x_2.shape[0] - 1
+        xx1 = np.c_[x_1[:-1], x_1[1:]]
+        xx2 = np.c_[x_2[:-1], x_2[1:]]
+        return (
+            np.less_equal(
+                np.tile(xx1.min(axis=1), (n2, 1)).T, np.tile(xx2.max(axis=1), (n1, 1))
+            ),
+            np.greater_equal(
+                np.tile(xx1.max(axis=1), (n2, 1)).T, np.tile(xx2.min(axis=1), (n1, 1))
+            ),
+        )
+
+    x_x = inner_inter(x1, x2)
+    z_z = inner_inter(y1, y2)
+    m, k = np.nonzero(x_x[0] & x_x[1] & z_z[0] & z_z[1])
+    n = len(m)
+    a_m, xz, b_m = np.zeros((4, 4, n)), np.zeros((4, n)), np.zeros((4, n))
+    a_m[0:2, 2, :] = -1
+    a_m[2:4, 3, :] = -1
+    a_m[0::2, 0, :] = np.diff(np.c_[x1, y1], axis=0)[m, :].T
+    a_m[1::2, 1, :] = np.diff(np.c_[x2, y2], axis=0)[k, :].T
+    b_m[0, :] = -x1[m].ravel()
+    b_m[1, :] = -x2[k].ravel()
+    b_m[2, :] = -y1[m].ravel()
+    b_m[3, :] = -y2[k].ravel()
+    for i in range(n):
+        try:
+            xz[:, i] = np.linalg.solve(a_m[:, :, i], b_m[:, i])
+        except np.linalg.LinAlgError:
+            # Parallel segments. Will raise numpy RuntimeWarnings
+            xz[0, i] = np.nan
+    in_range = (xz[0, :] >= 0) & (xz[1, :] >= 0) & (xz[0, :] <= 1) & (xz[1, :] <= 1)
+    xz = xz[2:, in_range].T
+    x, z = xz[:, 0], xz[:, 1]
+    if len(x) > 0:
+        x, z = np.unique([x, z], axis=1)
+    return x, z
+
+
+@nb.jit(cache=True, nopython=True)
+def _intersect_count(x_inter, z_inter, x2, z2):
+    args = []
+    for i in range(len(x_inter)):
+        for j in range(len(x2) - 1):
+            if check_linesegment(
+                np.array([x2[j], z2[j]]),
+                np.array([x2[j + 1], z2[j + 1]]),
+                np.array([x_inter[i], z_inter[i]]),
+            ):
+                args.append(j)
+                break
+    return np.array(args)
+
+
+def join_intersect(coords1, coords2, get_arg=False):
+    """
+    Add the intersection points between coords1 and coords2 to coords1.
+
+    Parameters
+    ----------
+    coords1: Coordinates
+        The Coordinates to which the intersection points should be added
+    coords2: Coordinates
+        The intersecting Coordinates
+    get_arg: bool (default = False)
+
+    Returns
+    -------
+    (if get_arg is True)
+    args: list(int, int, ..) of len(N_intersections)
+        The arguments of coords1 in which the intersections were added.
+
+    Notes
+    -----
+    Modifies coords1
+    """
+    x_inter, z_inter = get_intersect(coords1.xz, coords2.xz)
+    args = _intersect_count(x_inter, z_inter, coords1.x, coords1.z)
+
+    orderr = args.argsort()
+    x_int = x_inter[orderr]
+    z_int = z_inter[orderr]
+
+    args = _intersect_count(x_int, z_int, coords1.x, coords1.z)
+
+    # TODO: Check for duplicates and order correctly based on distance
+    # u, counts = np.unique(args, return_counts=True)
+
+    count = 0
+    for i, arg in enumerate(args):
+        if i > 0 and args[i - 1] == arg:
+            # Two intersection points, one after the other
+            bump = 0
+        else:
+            bump = 1
+        if not np.isclose(coords1.xyz.T, [x_int[i], 0, z_int[i]]).all(axis=1).any():
+            # Only increment counter if the intersection isn't already in the Coordinates
+            coords1.insert([x_int[i], 0, z_int[i]], index=arg + count + bump)
+            count += 1
+
+    if get_arg:
+        args = []
+        for x, z in zip(x_inter, z_inter):
+            args.append(coords1.argmin([x, 0, z]))
+        return list(set(args))
 
 
 # =============================================================================
